+++ notify-stack/src/main/java/com/dwarfeng/notify/stack/handler/DispatchLocalCacheHandler.java
@@ -1,8 +1,8 @@
 package com.dwarfeng.notify.stack.handler;
 
+import com.dwarfeng.notify.stack.bean.dto.DispatchInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.exception.HandlerException;
-import com.dwarfeng.subgrade.stack.handler.Handler;
+import com.dwarfeng.subgrade.stack.handler.LocalCacheHandler;
 
 /**
  * 调度本地缓存处理器。
@@ -14,30 +14,5 @@ import com.dwarfeng.subgrade.stack.handler.Handler;
  * @author DwArFeng
  * @since 1.0.0
  */
-public interface DispatchLocalCacheHandler extends Handler {
-
-    /**
-     * 获取指定主键对应的调度器。
-     *
-     * @param dispatcherInfoKey 指定的主键。
-     * @return 指定主键对应的调度器，如果不存在，则返回 null。
-     * @throws HandlerException 处理器异常。
-     */
-    Dispatcher getDispatcher(StringIdKey dispatcherInfoKey) throws HandlerException;
-
-    /**
-     * 获取指定主键对应的调度器类型。
-     *
-     * @param dispatcherInfoKey 指定的主键。
-     * @return 指定主键对应的调度器类型，如果不存在，则返回 null。
-     * @throws HandlerException 处理器异常。
-     */
-    String getType(StringIdKey dispatcherInfoKey) throws HandlerException;
-
-    /**
-     * 清除本地缓存。
-     *
-     * @throws HandlerException 处理器异常。
-     */
-    void clear() throws HandlerException;
+public interface DispatchLocalCacheHandler extends LocalCacheHandler<StringIdKey, DispatchInfo> {
 }